:memo: Correct scalar-TE docstrings and split the tie/no-fit selector test
Addresses review feedback on #32 and unblocks the lint job.

- `mcpc_3d_s` documents `te0`/`te1` as `np.float32 | float` to match the
  scalar signature; `rescale_phase` documents `min`/`max` as floats.
- Split `test_branch_selector_noop_when_nothing_fits` in two. The all-zero
  intercepts case is an all-way perfect fit that collapses the observed scale,
  not a failed fit, so it gets its own test and an accurate docstring.
- Reformat the README's Python block: ruff 0.16 formats code in markdown, which
  fails `ruff format --check` on main as well as here.

+++ warpkit/utilities.py
@@ -151,9 +151,9 @@ def rescale_phase(
     ----------
     data : npt.NDArray[Any]
         phase data to be rescaled
-    min : int, optional
+    min : float, optional
         min value that should be mapped to -pi, by default -4096
-    max : int, optional
+    max : float, optional
         max value that should be mapped to pi, by default 4096
 
     Returns
